move generated files
