+++ src/grpc/node.ts
@@ -3,16 +3,16 @@ import { Utils } from 'xpring-common-js'
 import {
   GetAccountInfoRequest,
   GetAccountInfoResponse,
-} from '../generated/node/rpc/v1/account_info_pb'
-import { GetFeeRequest, GetFeeResponse } from '../generated/node/rpc/v1/fee_pb'
-import { GetTxRequest, GetTxResponse } from '../generated/node/rpc/v1/tx_pb'
+} from './generated/node/rpc/v1/account_info_pb'
+import { GetFeeRequest, GetFeeResponse } from './generated/node/rpc/v1/fee_pb'
+import { GetTxRequest, GetTxResponse } from './generated/node/rpc/v1/tx_pb'
 import {
   SubmitTransactionRequest,
   SubmitTransactionResponse,
-} from '../generated/node/rpc/v1/submit_pb'
-import { XRPLedgerAPIServiceClient } from '../generated/node/rpc/v1/xrp_ledger_grpc_pb'
+} from './generated/node/rpc/v1/submit_pb'
+import { XRPLedgerAPIServiceClient } from './generated/node/rpc/v1/xrp_ledger_grpc_pb'
 import { NetworkClientDecorator } from '../utils/network-client-decorator'
-import { AccountAddress } from '../generated/node/rpc/v1/amount_pb'
+import { AccountAddress } from './generated/node/rpc/v1/amount_pb'
 import isNode from '../utils'
 
 /**
