+++ src/legacy/grpc/node.ts
@@ -1,20 +1,20 @@
 import { credentials } from 'grpc'
 import { Wallet, Signer } from 'xpring-common-js'
-import { XRPLedgerAPIClient } from '../../generated/node/legacy/xrp_ledger_grpc_pb'
-import { AccountInfo } from '../../generated/node/legacy/account_info_pb'
-import { Fee } from '../../generated/node/legacy/fee_pb'
-import { GetFeeRequest } from '../../generated/node/legacy/get_fee_request_pb'
-import { GetAccountInfoRequest } from '../../generated/node/legacy/get_account_info_request_pb'
-import { SubmitSignedTransactionRequest } from '../../generated/node/legacy/submit_signed_transaction_request_pb'
-import { SubmitSignedTransactionResponse } from '../../generated/node/legacy/submit_signed_transaction_response_pb'
-import { GetLatestValidatedLedgerSequenceRequest } from '../../generated/node/legacy/get_latest_validated_ledger_sequence_request_pb'
-import { LedgerSequence } from '../../generated/node/legacy/ledger_sequence_pb'
-import { GetTransactionStatusRequest } from '../../generated/node/legacy/get_transaction_status_request_pb'
-import { TransactionStatus } from '../../generated/node/legacy/transaction_status_pb'
-import { XRPAmount } from '../../generated/node/legacy/xrp_amount_pb'
-import { Payment } from '../../generated/node/legacy/payment_pb'
-import { Transaction } from '../../generated/node/legacy/transaction_pb'
-import { SignedTransaction } from '../../generated/node/legacy/signed_transaction_pb'
+import { XRPLedgerAPIClient } from './generated/node/xrp_ledger_grpc_pb'
+import { AccountInfo } from './generated/node/account_info_pb'
+import { Fee } from './generated/node/fee_pb'
+import { GetFeeRequest } from './generated/node/get_fee_request_pb'
+import { GetAccountInfoRequest } from './generated/node/get_account_info_request_pb'
+import { SubmitSignedTransactionRequest } from './generated/node/submit_signed_transaction_request_pb'
+import { SubmitSignedTransactionResponse } from './generated/node/submit_signed_transaction_response_pb'
+import { GetLatestValidatedLedgerSequenceRequest } from './generated/node/get_latest_validated_ledger_sequence_request_pb'
+import { LedgerSequence } from './generated/node/ledger_sequence_pb'
+import { GetTransactionStatusRequest } from './generated/node/get_transaction_status_request_pb'
+import { TransactionStatus } from './generated/node/transaction_status_pb'
+import { XRPAmount } from './generated/node/xrp_amount_pb'
+import { Payment } from './generated/node/payment_pb'
+import { Transaction } from './generated/node/transaction_pb'
+import { SignedTransaction } from './generated/node/signed_transaction_pb'
 import { LegacyNetworkClient } from './network-client'
 import XpringClientErrorMessages from '../../utils/xpring-client-error-messages'
