+++ src/legacy/grpc/web.ts
@@ -1,20 +1,20 @@
 import { Signer, Wallet } from 'xpring-common-js'
-import { XRPLedgerAPIClient } from '../../generated/web/legacy/xrp_ledger_grpc_web_pb'
-import { AccountInfo } from '../../generated/web/legacy/account_info_pb'
-import { Fee } from '../../generated/web/legacy/fee_pb'
-import { GetFeeRequest } from '../../generated/web/legacy/get_fee_request_pb'
-import { GetAccountInfoRequest } from '../../generated/web/legacy/get_account_info_request_pb'
-import { SubmitSignedTransactionRequest } from '../../generated/web/legacy/submit_signed_transaction_request_pb'
-import { SubmitSignedTransactionResponse } from '../../generated/web/legacy/submit_signed_transaction_response_pb'
-import { GetLatestValidatedLedgerSequenceRequest } from '../../generated/web/legacy/get_latest_validated_ledger_sequence_request_pb'
-import { LedgerSequence } from '../../generated/web/legacy/ledger_sequence_pb'
-import { GetTransactionStatusRequest } from '../../generated/web/legacy/get_transaction_status_request_pb'
-import { TransactionStatus } from '../../generated/web/legacy/transaction_status_pb'
-import { XRPAmount } from '../../generated/node/legacy/xrp_amount_pb'
-import { Payment } from '../../generated/node/legacy/payment_pb'
-import { Transaction } from '../../generated/node/legacy/transaction_pb'
+import { XRPLedgerAPIClient } from './generated/web/xrp_ledger_grpc_web_pb'
+import { AccountInfo } from './generated/web/account_info_pb'
+import { Fee } from './generated/web/fee_pb'
+import { GetFeeRequest } from './generated/web/get_fee_request_pb'
+import { GetAccountInfoRequest } from './generated/web/get_account_info_request_pb'
+import { SubmitSignedTransactionRequest } from './generated/web/submit_signed_transaction_request_pb'
+import { SubmitSignedTransactionResponse } from './generated/web/submit_signed_transaction_response_pb'
+import { GetLatestValidatedLedgerSequenceRequest } from './generated/web/get_latest_validated_ledger_sequence_request_pb'
+import { LedgerSequence } from './generated/web/ledger_sequence_pb'
+import { GetTransactionStatusRequest } from './generated/web/get_transaction_status_request_pb'
+import { TransactionStatus } from './generated/web/transaction_status_pb'
+import { XRPAmount } from './generated/node/xrp_amount_pb'
+import { Payment } from './generated/node/payment_pb'
+import { Transaction } from './generated/node/transaction_pb'
 import isNode from '../../utils'
-import { SignedTransaction } from '../../generated/web/legacy/signed_transaction_pb'
+import { SignedTransaction } from './generated/web/signed_transaction_pb'
 import { LegacyNetworkClient } from './network-client'
 import XpringClientErrorMessages from '../../utils/xpring-client-error-messages'
