+++ src/utils/network-client-decorator.ts
@@ -1,10 +1,10 @@
-import { GetAccountInfoResponse } from '../generated/web/rpc/v1/account_info_pb'
-import { GetFeeResponse } from '../generated/web/rpc/v1/fee_pb'
-import { GetTxResponse } from '../generated/web/rpc/v1/tx_pb'
+import { GetAccountInfoResponse } from '../grpc/generated/web/rpc/v1/account_info_pb'
+import { GetFeeResponse } from '../grpc/generated/web/rpc/v1/fee_pb'
+import { GetTxResponse } from '../grpc/generated/web/rpc/v1/tx_pb'
 import {
   SubmitTransactionRequest,
   SubmitTransactionResponse,
-} from '../generated/web/rpc/v1/submit_pb'
+} from '../grpc/generated/web/rpc/v1/submit_pb'
 
 /**
  * The network client interface provides a wrapper around network calls to the Xpring Platform.
